Moving SensorThings code to ST library
